+++ libs/base_core/src/ui/front/term/ansi/mod.rs
@@ -8,7 +8,7 @@
 #![expect(non_snake_case, reason = "uppercase const fns as related constants")]
 
 mod namespace; // Ansi
-mod color; // AnsiColor3b, AnsiColor8b (+Ansi impls)
+mod color; // AnsiColor3, AnsiColor8 (+Ansi impls)
 
 crate::structural_mods! { // _mods
     _mods {